Move snapshot class to dataset module
